Moved spawner into dedicated class of its own
